+++ vscode-extension/src/lib/AppIdCache.ts
@@ -1,11 +1,11 @@
 import { PropertyBag } from "./PropertyBag";
-import * as crypto from "crypto";
+import { getSha256 } from "./Sha256";
 
 export class AppIdCache {
     private _map: PropertyBag<string> = {};
     private static _instance: AppIdCache;
 
-    private constructor() {}
+    private constructor() { }
 
     public static get instance(): AppIdCache {
         return this._instance || (this._instance = new AppIdCache());
@@ -13,9 +13,6 @@ export class AppIdCache {
 
     public getAppIdHash(appId: string): string {
         if (this._map[appId]) return this._map[appId];
-
-        const sha256 = crypto.createHash("sha256");
-        sha256.update(appId);
-        return this._map[appId] = sha256.digest("hex");
+        return this._map[appId] = getSha256(appId);
     }
 }